# New Span-based C# parsing logic