@jtattermusch 2020-04-02
CC @JamesNK @JunTaoLuo 

this is still WIP and I'll probably do some rebasing, but all this seems to be the minimal PR that adds ReadOnlySequence parsing.