@JamesNK 2020-04-11
Changes to fix running on .NET Framework:

https://github.com/JamesNK/protobuf/commit/5d868ad9f34c0f850ff0ba450d75370d1cb9a8cf