@jtattermusch 2020-04-16
ParseMessagesBenchmark results:

The original _ParseFromByteArray performance is basically the same as the new __ParseFromReadOnlySequence performance, except that initializing the ParseContext is ~40ns slower than initializing the original CodedInputSteam, so there is a one-time overhead when parsing a single message (one can best see this in the parse EmptyMessage benchmark). This overhead seems completely negligible whenever we start parsing non-trivial messages and such small overhead also can't be noticed in e.g. RPC scenarios where the overhead of an RPC call will be many times larger than this.

The new _ParseFromByteArray scenario still uses CodedInputStream for parsing the messages, so there's extra overhead (once per message parsed) converting the CodedInputStream's internal state into a parseContext (which requires a struct copy). Under normal circumstances this overhead will only be incurred at the root message of the parsing tree and it's unlikely to matter in real-life scenarios.

```
BEFORE

BenchmarkDotNet=v0.11.4, OS=macOS Mojave 10.14.6 (18G3020) [Darwin 18.7.0]
Intel Core i7-7567U CPU 3.50GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=3.0.100
  [Host]     : .NET Core 2.1.15 (CoreCLR 4.6.28325.01, CoreFX 4.6.28327.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.15 (CoreCLR 4.6.28325.01, CoreFX 4.6.28327.02), 64bit RyuJIT


|                                                         Method | messageCount |         Mean |       Error |      StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|--------------------------------------------------------------- |------------- |-------------:|------------:|------------:|------------:|------------:|------------:|--------------------:|
|                    ManyWrapperFieldsMessage_ParseFromByteArray |            ? |    381.41 ns |   8.6719 ns |  24.3170 ns |      0.8879 |           - |           - |              1864 B |
|                  ManyPrimitiveFieldsMessage_ParseFromByteArray |            ? |    186.74 ns |   2.5298 ns |   2.2426 ns |      0.4880 |           - |           - |              1024 B |
|                                EmptyMessage_ParseFromByteArray |            ? |     37.50 ns |   0.6567 ns |   0.5821 ns |      0.0533 |           - |           - |               112 B |
|   ManyWrapperFieldsMessage_ParseDelimitedMessagesFromByteArray |           10 |  3,405.47 ns |  48.4968 ns |  40.4970 ns |      8.5106 |           - |           - |             17848 B |
| ManyPrimitiveFieldsMessage_ParseDelimitedMessagesFromByteArray |           10 |  1,799.61 ns |  22.4263 ns |  20.9775 ns |      4.5033 |           - |           - |              9448 B |
|   ManyWrapperFieldsMessage_ParseDelimitedMessagesFromByteArray |          100 | 33,946.81 ns | 444.3343 ns | 393.8908 ns |     84.6558 |           - |           - |            177688 B |
| ManyPrimitiveFieldsMessage_ParseDelimitedMessagesFromByteArray |          100 | 18,146.69 ns | 356.7617 ns | 424.6995 ns |     44.6472 |           - |           - |             93688 B |

AFTER

BenchmarkDotNet=v0.11.4, OS=macOS Mojave 10.14.6 (18G3020) [Darwin 18.7.0]
Intel Core i7-7567U CPU 3.50GHz (Kaby Lake), 1 CPU, 4 logical and 2 physical cores
.NET Core SDK=3.0.100
  [Host]     : .NET Core 2.1.15 (CoreCLR 4.6.28325.01, CoreFX 4.6.28327.02), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.15 (CoreCLR 4.6.28325.01, CoreFX 4.6.28327.02), 64bit RyuJIT


|                                                                Method | messageCount |         Mean |       Error |      StdDev | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|---------------------------------------------------------------------- |------------- |-------------:|------------:|------------:|------------:|------------:|------------:|--------------------:|
|                           ManyWrapperFieldsMessage_ParseFromByteArray |            ? |    409.88 ns |   7.4969 ns |   6.6458 ns |      0.9341 |           - |           - |              1960 B |
|                    ManyWrapperFieldsMessage_ParseFromReadOnlySequence |            ? |    381.03 ns |   8.7381 ns |   8.1736 ns |      0.8464 |           - |           - |              1776 B |
|                         ManyPrimitiveFieldsMessage_ParseFromByteArray |            ? |    251.87 ns |   4.6516 ns |   4.1235 ns |      0.5336 |           - |           - |              1120 B |
|                  ManyPrimitiveFieldsMessage_ParseFromReadOnlySequence |            ? |    223.52 ns |   4.4897 ns |   6.4390 ns |      0.4461 |           - |           - |               936 B |
|                                       EmptyMessage_ParseFromByteArray |            ? |    106.58 ns |   0.9453 ns |   0.8380 ns |      0.0991 |           - |           - |               208 B |
|                                EmptyMessage_ParseFromReadOnlySequence |            ? |     71.74 ns |   1.0780 ns |   1.0084 ns |      0.0113 |           - |           - |                24 B |
|          ManyWrapperFieldsMessage_ParseDelimitedMessagesFromByteArray |           10 |  3,906.78 ns |  41.3674 ns |  36.6711 ns |      8.5526 |           - |           - |             17944 B |
|   ManyWrapperFieldsMessage_ParseDelimitedMessagesFromReadOnlySequence |           10 |  3,483.34 ns |  39.6573 ns |  37.0955 ns |      8.4648 |           - |           - |             17760 B |
|        ManyPrimitiveFieldsMessage_ParseDelimitedMessagesFromByteArray |           10 |  2,235.63 ns |  21.6903 ns |  19.2279 ns |      4.5471 |           - |           - |              9544 B |
| ManyPrimitiveFieldsMessage_ParseDelimitedMessagesFromReadOnlySequence |           10 |  1,820.10 ns |  25.1990 ns |  22.3383 ns |      4.4613 |           - |           - |              9360 B |
|          ManyWrapperFieldsMessage_ParseDelimitedMessagesFromByteArray |          100 | 39,119.34 ns | 493.0260 ns | 461.1768 ns |     84.7168 |           - |           - |            177784 B |
|   ManyWrapperFieldsMessage_ParseDelimitedMessagesFromReadOnlySequence |          100 | 36,101.14 ns | 678.6854 ns | 634.8427 ns |     84.6558 |           - |           - |            177600 B |
|        ManyPrimitiveFieldsMessage_ParseDelimitedMessagesFromByteArray |          100 | 22,013.79 ns | 278.3219 ns | 260.3425 ns |     44.7083 |           - |           - |             93784 B |
| ManyPrimitiveFieldsMessage_ParseDelimitedMessagesFromReadOnlySequence |          100 | 18,053.55 ns | 299.3745 ns | 280.0351 ns |     44.6167 |           - |           - |             93600 B |
```